added useTagsHandler custom hook to consolidate getTagColor and presetOptions state into one function
